refactor: removing marked to use our own parser
This commit removes marked in order to use our own parser, guaranteed
with no regex, as Discord's markdown implementation is weird; in that
way, there are a lot of new use cases to support, which marked does
not.

This commit also update dependencies and refactor a bit more the code.