refactor: better error handling (#5)
* refactor: better error handling

This commit adds better error handling. Now, it no longer deletes
original messages if any error had occurred.

Errors coming from the bin service are also formatted to make them
clearer for people.

* style: eslint

* Update src/misc/BinError.ts

Co-authored-by: Mestery <48163546+Mesteery@users.noreply.github.com>

* Update src/helpers/index.ts

Co-authored-by: Mestery <48163546+Mesteery@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: Mestery <48163546+Mesteery@users.noreply.github.com>

* style, code correction, etc etc

* semi-support for multiple attachments

* max size updated to 8MiB instead of 8MB

* update max size

* a bit of refactor

* fix issue with total size

* blabla

* continue instead of break

* add size check

* fix: log criticals errors

This commit allows to log the error if it is criterical.
In addition, the error was not reported at all if there were more errors than changes. So there may be cases, or the error is not reported, even if it is critical.

* Update src/helpers/contentProcessing.ts

Co-authored-by: Mestery <48163546+Mesteery@users.noreply.github.com>

* refactor: add error logger helper

Co-authored-by: Mestery <48163546+Mesteery@users.noreply.github.com>