feat: add bot ping (#21)
* chore(deps): update dependencies

This commit updates dependencies. It also adds @types/jest as a dev
dependency since test tools seem apparently to have splitted their types
definitions.

* fix(types): typing problem

This commit fixes a typing problem as TypeScript did not expect a
non-const options parameter to go with Date#toLocaleDateString.

* style: remove useless variable

* feat: add ping feature

This commit adds a ping feature to the bot. All we have to do is to ping
it, in any channel (even if not in a correct category).

* refactor: change ping feature

This commit changes how the ping feature is implemented. It now uses an
embed and a command-like system by using the bot mention as a prefix
followed by the "ping" command.

* refactor: change how bin's status is showed

* refactor: change red color

* refactor: change green

* refactor: use Array#includes

* refactor: idk how to name it please send help