feat: transform message when requested
When the message's author add :floppy_disk: reaction
to his message, the bot will process the message
even if the max lines is not exceeded. This will
allows to explicity request the bot to transform
a message.

This is also possible if there are more than
'TRANSFORM_MIN_REACTIONS' reactions to the message
or if the member who reacted has the 'MANAGE_MESSAGES'
permission.